+++ src/main/resources/query/CREATE_TABLE.sql
@@ -1,8 +1,8 @@
-CREATE TABLE student_group (id INTEGER NOT NULL,
+CREATE TABLE student_group (id BIGINT NOT NULL,
                     number INTEGER DEFAULT NULL,
                     faculty VARCHAR(200) DEFAULT NULL,
                     PRIMARY KEY (id));
-CREATE TABLE student (id INTEGER NOT NULL,
+CREATE TABLE student (id BIGINT NOT NULL,
                       firstname VARCHAR(50) DEFAULT NULL,
                       lastname VARCHAR(50) DEFAULT NULL,
                       secondname VARCHAR(50) DEFAULT NULL,
